Limpeza de codigo e padronizacao de excecoes

+++ src/main/java/br/com/amparo/backend/service/impl/DosageServiceImpl.java
@@ -3,8 +3,11 @@ package br.com.amparo.backend.service.impl;
 import br.com.amparo.backend.dto.dosage.AddDosageRequest;
 import br.com.amparo.backend.dto.dosage.DosageResponse;
 import br.com.amparo.backend.dto.dosage.EditDosageRequest;
+import br.com.amparo.backend.exception.PatientNotFoundException;
 import br.com.amparo.backend.repository.DosageRepository;
 import br.com.amparo.backend.service.DosageService;
+import br.com.amparo.backend.service.PatientService;
+import br.com.amparo.backend.service.security.SecurityUtils;
 import lombok.RequiredArgsConstructor;
 
 import java.util.Optional;
@@ -12,23 +15,49 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class DosageServiceImpl implements DosageService {
     private final DosageRepository repository;
+    private final PatientService patientService;
     @Override
-    public Optional<DosageResponse> addDosage(String patientId, String medicineId, AddDosageRequest request) {
-        return repository.addDosage(patientId, medicineId, request);
+    public Optional<DosageResponse> addDosage(String medicineId, AddDosageRequest request) {
+        String patientId = SecurityUtils.getApiUser().getId();
+        if (patientService.findPatientById(patientId).isEmpty()) {
+            throw new PatientNotFoundException(patientId);
+        } else {
+            return repository.addDosage(patientId, medicineId, request);
+        }
+    }
+    @Override
+    public Optional<DosageResponse> getDosage(String dosageId) throws IllegalAccessException {
+        String patientId = SecurityUtils.getApiUser().getId();
+        if (patientService.findPatientById(patientId).isEmpty()) {
+            throw new PatientNotFoundException(patientId);
+        } else if (!patientId.equals(repository.getDosage(dosageId).get().idPatient())) {
+            throw new IllegalAccessException(patientId);
+        } else {
+            return repository.getDosage(dosageId);
+        }
     }
 
     @Override
-    public Optional<DosageResponse> getDosage(String dosageId) {
-        return repository.getDosage(dosageId);
+    public Optional<DosageResponse> deleteDosage(DosageResponse dosage) throws IllegalAccessException {
+        String patientId = SecurityUtils.getApiUser().getId();
+        if (patientService.findPatientById(patientId).isEmpty()) {
+            throw new PatientNotFoundException(patientId);
+        } else if (!patientId.equals(repository.getDosage(dosage.id()).get().idPatient())) {
+            throw new IllegalAccessException(patientId);
+        } else {
+            return repository.deleteDosage(dosage);
+        }
     }
 
     @Override
-    public Optional<DosageResponse> deleteDosage(DosageResponse dosage) {
-        return repository.deleteDosage(dosage);
-    }
-
-    @Override
-    public Optional<DosageResponse> editDosage(String patientId, String dosageId, EditDosageRequest request) {
-        return repository.editDosage(dosageId, request);
+    public Optional<DosageResponse> editDosage(String dosageId, EditDosageRequest request) throws IllegalAccessException {
+        String patientId = SecurityUtils.getApiUser().getId();
+        if (patientService.findPatientById(patientId).isEmpty()) {
+            throw new PatientNotFoundException(patientId);
+        } else if (!patientId.equals(repository.getDosage(dosageId).get().idPatient())) {
+            throw new IllegalAccessException(patientId);
+        } else {
+            return repository.editDosage(dosageId, request);
+        }
     }
 }
